chore: cleanup codes and add download link

+++ source/Scores/SaveScores.cpp
@@ -9,7 +9,7 @@
 void saveScore(std::vector<Score>& scores) {
     std::ofstream file;
     file.open("scores.csv", std::ios::out);
-    for (int i = 0; i < scores.size(); i++) {
+    for (unsigned int i = 0; i < (unsigned int)scores.size(); i++) {
         file << scores[i].getName() << "," << scores[i].getScore() << std::endl;
     }
     file.close();
